+++ nucliadb_vectors/Cargo.toml
@@ -22,6 +22,7 @@ rayon = "1.5.2"
 memmap2 = "0.5.3"
 fs2 = "0.4.3"
 nucliadb_service_interface = { path = "../nucliadb_service_interface" }
+nucliadb_byte_rpr = { path = "../nucliadb_byte_rpr" }
 clap = { version = "3.1.18", features = ["derive"] }
 [dev-dependencies]
 env_logger = "0.9.0"
