+++ nucliadb_vectors/src/memory_system/elements/definitions.rs
@@ -1,13 +1,28 @@
+// Copyright (C) 2021 Bosutech XXI S.L.
+//
+// nucliadb is offered under the AGPL v3.0 and as commercial software.
+// For commercial licensing, contact us at info@nuclia.com.
+//
+// AGPL:
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+//
+
 use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
 
 pub const VECTORS_DIR: &str = "vectors";
-
-pub trait ByteRpr {
-    fn as_byte_rpr(&self) -> Vec<u8>;
-    fn from_byte_rpr(bytes: &[u8]) -> Self;
-}
 
 pub trait Distance {
     fn cosine(i: &Self, j: &Self) -> f32;
